HipSpecifyer.Prelude

Documentation

genericArbitrary :: forall a. (Typeable a, Arbitrary a, Generic a, GArbitrary (Rep a)) => Gen a

genericCoarbitrary :: (Generic a, GCoarbitrary (Rep a)) => a -> Gen b -> Gen b

class Observe a where

Methods

observe :: a -> Gen Observation

genericObserve :: (Generic a, GObserve (Rep a)) => a -> Gen Observation

obs0 :: (Observe a, Typeable a) => String -> a -> Sig

obs1 :: (Typeable a, Typeable b, Observe b) => String -> (a -> b) -> Sig

A unary function.

obs2 :: (Typeable a, Typeable b, Typeable c, Observe c) => String -> (a -> b -> c) -> Sig

A binary function.

obs3 :: (Typeable a, Typeable b, Typeable c, Typeable d, Observe d) => String -> (a -> b -> c -> d) -> Sig

A ternary function.

obs4 :: (Typeable a, Typeable b, Typeable c, Typeable d, Typeable e, Observe e) => String -> (a -> b -> c -> d -> e) -> Sig

A function of four arguments.

obs5 :: (Typeable a, Typeable b, Typeable c, Typeable d, Typeable e, Typeable f, Observe f) => String -> (a -> b -> c -> d -> e -> f) -> Sig

A function of five arguments.